Frontend added

+++ src/app.py
@@ -1,5 +1,5 @@
 from fastapi import FastAPI,HTTPException,File,UploadFile,Form,Depends
-from .db.db import create_db_and_tables,get_session,Post
+from .db.db import create_db_and_tables,get_session,Post,User
 from sqlalchemy.ext.asyncio import AsyncSession
 from contextlib import asynccontextmanager
 from sqlalchemy.future import select
@@ -31,6 +31,7 @@ async def read_root():
 async def upload_file(
         file: UploadFile = File(...),
         caption: str = Form(""),
+        user: User = Depends(current_active_user),
         session: AsyncSession = Depends(get_session)
 ):
     file_content = await file.read()
@@ -49,6 +50,7 @@ async def upload_file(
     else:
         file_type = "file"
     post = Post(
+        user_id=user.id,
         caption=caption,
         url=file_url,
         file_type=file_type,
@@ -61,30 +63,36 @@ async def upload_file(
 
 
 @app.get("/feed")
-async def get_feed(session: AsyncSession = Depends(get_session)):
+async def get_feed(session: AsyncSession = Depends(get_session),user: User = Depends(current_active_user)):
     result = await session.execute(select(Post).order_by(Post.created_at.desc()))
     posts = [row[0] for row in result.all()]
+    
     
     posts_data = []
     for post in posts:
         posts_data.append({
             "id": str(post.id),
+            "user_id": str(post.user_id),
             "caption": post.caption,
             "url": post.url,
             "file_type": post.file_type,
             "file_name": post.file_name,
-            "created_at": post.created_at.isoformat()
+            "created_at": post.created_at.isoformat(),
+            "is_owner": post.user_id == user.id ,
+            "email": user.email
         })
     return posts_data
 
 @app.delete("/posts/{post_id}")
-async def delete_post(post_id: str, session: AsyncSession = Depends(get_session)):
+async def delete_post(post_id: str, session: AsyncSession = Depends(get_session),user: User = Depends(current_active_user)):
     try:
         post_id = uuid.UUID(post_id)
         result = await session.execute(select(Post).where(Post.id == post_id))
         post = result.scalar_one_or_none()
         if not post:
             raise HTTPException(status_code=404, detail="Post not found")
+        if post.user_id != user.id:
+            raise HTTPException(status_code=403, detail="Not authorized to delete this post")
         await session.delete(post)
         await session.commit()
         return {"detail": "Post deleted successfully"}
